fix: correção da tipagem int para Long

+++ educa100/src/main/java/com/educa100/controller/dto/request/DocenteRequest.java
@@ -2,5 +2,5 @@ package com.educa100.controller.dto.request;
 
 import java.util.Date;
 
-public record DocenteRequest(String nome, Date dataEntrada, int id_usuario) {
+public record DocenteRequest(String nome, Date dataEntrada, Long id_usuario) {
 }
