+++ educa100/src/main/java/com/educa100/controller/dto/request/MateriaRequest.java
@@ -1,4 +1,4 @@
 package com.educa100.controller.dto.request;
 
-public record MateriaRequest(String nome, int id_curso) {
+public record MateriaRequest(String nome, Long id_curso) {
 }
