+++ educa100/src/main/java/com/educa100/infra/exception/MateriaNotFoundException.java
@@ -1,7 +1,7 @@
 package com.educa100.infra.exception;
 
 public class MateriaNotFoundException extends NotFoundException{
-    public MateriaNotFoundException(int id){
+    public MateriaNotFoundException(Long id){
         super("Materia com o id " +id +" não encontrado");
     }
 }
